refactor(Rng): Prelude module, code reorg & State made private
