+++ src/prelude.rs
@@ -1,0 +1,15 @@
+//! Convenience re-export of common traits, structs and utils.
+
+pub use crate::traits::*;
+
+#[cfg(any(feature = "wyrand", feature = "atomic"))]
+#[cfg_attr(docsrs, doc(cfg(any(feature = "wyrand", feature = "atomic"))))]
+pub use crate::{internal::*, rng::*};
+
+#[cfg(feature = "chacha")]
+#[cfg_attr(docsrs, doc(cfg(feature = "chacha")))]
+pub use crate::secure_rng::*;
+
+#[cfg(feature = "rand")]
+#[cfg_attr(docsrs, doc(cfg(feature = "rand")))]
+pub use crate::compatibility::*;
